requirements.txt
tensorflow>=2.10
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.5
pillow>=9.0
PyYAML>=6.0
pytest>=7.0
